+++ Borealis/src/Graphics/RenderGraph/RenderGraph.cpp
@@ -935,7 +935,7 @@ namespace Borealis
 			{
 				TextComponent const& text = parent.GetComponent<TextComponent>();
 
-				Renderer2D::DrawString(text.text, text.font, transform, (int)parent, text.fontSize, text.colour, text.align == TextComponent::TextAlign::Left ? false : true, text.outline, text.width);
+				Renderer2D::DrawString(text.text, text.font, transform, (int)parent, (float)text.fontSize, text.colour, text.align == TextComponent::TextAlign::Left ? false : true, text.outline, text.width);
 			}
 
 			if (parent.HasComponent<UIAnimatorComponent>())
